+++ cmd/parse/cmd.go
@@ -3,11 +3,11 @@ package parse
 import (
 	"github.com/spf13/cobra"
 
-	parsecmdtypes "github.com/forbole/juno/v3/cmd/parse/types"
+	parsecmdtypes "github.com/forbole/juno/v4/cmd/parse/types"
 
-	parseblocks "github.com/forbole/juno/v3/cmd/parse/blocks"
-	parsegenesis "github.com/forbole/juno/v3/cmd/parse/genesis"
-	parsetransactions "github.com/forbole/juno/v3/cmd/parse/transactions"
+	parseblocks "github.com/forbole/juno/v4/cmd/parse/blocks"
+	parsegenesis "github.com/forbole/juno/v4/cmd/parse/genesis"
+	parsetransactions "github.com/forbole/juno/v4/cmd/parse/transactions"
 )
 
 // NewParseCmd returns the Cobra command allowing to parse some chain data without having to re-sync the whole database