+++ database/postgresql/postgresql_test.go
@@ -11,10 +11,10 @@ import (
 	"github.com/cosmos/cosmos-sdk/simapp"
 	"github.com/stretchr/testify/suite"
 
-	"github.com/forbole/juno/v3/database"
-	databaseconfig "github.com/forbole/juno/v3/database/config"
-	postgres "github.com/forbole/juno/v3/database/postgresql"
-	"github.com/forbole/juno/v3/logging"
+	"github.com/forbole/juno/v4/database"
+	databaseconfig "github.com/forbole/juno/v4/database/config"
+	postgres "github.com/forbole/juno/v4/database/postgresql"
+	"github.com/forbole/juno/v4/logging"
 )
 
 func TestDatabaseTestSuite(t *testing.T) {
@@ -33,13 +33,7 @@ func (suite *DbTestSuite) SetupTest() {
 
 	// Build the database
 	dbCfg := databaseconfig.NewDatabaseConfig(
-		"bdjuno",
-		"localhost",
-		6433,
-		"bdjuno",
-		"password",
-		"",
-		"public",
+		"postgres://bdjuno:password@localhost:6433/bdjuno?sslmode=disable&search_path=public",
 		-1,
 		-1,
 		100000,