+++ modules/messages/module.go
@@ -3,8 +3,8 @@ package messages
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
 
-	"github.com/forbole/juno/v3/database"
-	"github.com/forbole/juno/v3/modules"
+	"github.com/forbole/juno/v4/database"
+	"github.com/forbole/juno/v4/modules"
 )
 
 var (