+++ CHANGELOG.md
@@ -1,9 +1,18 @@
-## Unreleased
+## v4.1.0
 ### Changes
+- ([\#74](https://github.com/forbole/juno/pull/74)) Applied `GetMissingHeights()` in `enqueueMissingBlocks()` & in `parse blocks missing` cmd
+- ([\#88](https://github.com/forbole/juno/pull/88)) Added `juno_db_latest_height` metric to prometheus monitoring
+
+## v4.0.0
+### Changes
+- Updated cosmos/cosmos-sdk to `v0.45.8`
 - ([\#74](https://github.com/forbole/juno/pull/74)) Added database block count to prometheus to improve alert monitoring
 - ([\#75](https://github.com/forbole/juno/pull/75)) Allow modules to handle MsgExec inner messages
-- ([\#76](https://github.com/forbole/juno/pull/76)) Return 0 as height for `GetLastBlockHeight()` method while no block is saved
-- ([\#79](https://github.com/forbole/juno/pull/79)) Use `sqlx` instead of `sql` inside `PostgreSQLDatabase`
+- ([\#76](https://github.com/forbole/juno/pull/76)) Return 0 as height for `GetLastBlockHeight()` method if there are no blocks saved in database
+- ([\#79](https://github.com/forbole/juno/pull/79)) Use `sqlx` instead of `sql` while dealing with a PostgreSQL database
+- ([\#83](https://github.com/forbole/juno/pull/83)) Bump `github.com/tendermint/tendermint` to `v0.34.22`
+- ([\#84](https://github.com/forbole/juno/pull/84)) Replace database configuration params with URI
+- ([\#86](https://github.com/forbole/juno/pull/86)) Revert concurrent handling of transactions and messages
 
 ## v3.4.0
 ### Changes